Stop querying for the organization list.
The changes to the repositories and organization APIs are live
(https://developer.github.com/changes/2015-06-24-breaking-changes-to-organization-permissions-are-now-official/)
and thus the scopes used for the public repositories only option are no longer
sufficient to list organizations. However, it's no longer necessary to do that
in the first place, since the user repository list now includes repos from
organizations.